Initial work to get a restful API put together for the Code for Orlando's National Day of Civic Hacking.
This project is code-named: "Scott's App".

Scott works for City of Orlando and has to publish road closure data every Thursday.  He hand types it each and every time and then sends a PDF out to various consumers.  Would it not be cool if he could just enter each event and road closure into a database and have it auto-publish?  Yeah, that would be cool!"